backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
bcrypt
python-jose[cryptography]
pydantic
python-dotenv
alembic